requirements.txt
django==1.9.7 # the web server
mailsnake==1.6
# fabric==1.12.0 doesn't support python 3.... losers...
django-suit==0.2.21 # Admin theme
djangorestframework==3.5.3
Fabric3==1.12.0.post1
